Create common login logic
Prepare to mocking login feature
